Debugging Support Adds two new commands to language settings `debugCommand` and `debugAttachConfig`. The extension launches the `debugCommand` which typically involves invoking the debugger, and VSCode then attaches the debugger to that process.
Attach is used because we need to pass in the testcase inputs into it. We pass the input into the debug command (which the program will typically inherit its IO from the debugger process).

This is debugger-agnostic in the sense the user specifies how to invoke the debugger and attach to it.